Generalize the prompting and validation part
